+++ migrations/versions/0f594b84fbe7_.py
@@ -1,0 +1,24 @@
+"""empty message
+
+Revision ID: 0f594b84fbe7
+Revises: 2c52a938f5cc, c4e8e24d4a98
+Create Date: 2024-07-18 19:06:38.337919
+
+"""
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision = '0f594b84fbe7'
+down_revision = ('2c52a938f5cc', 'c4e8e24d4a98')
+branch_labels = None
+depends_on = None
+
+
+def upgrade() -> None:
+    pass
+
+
+def downgrade() -> None:
+    pass